+++ POSTerminalGui/Product.cs
@@ -9,11 +9,11 @@ namespace POSTerminalGui
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public Category ProductCategory { get; set; }
         public bool Taxable { get; set; }
 
-        public Product(string name, string description, double price, Category productCategory, bool taxable)
+        public Product(string name, string description, decimal price, Category productCategory, bool taxable)
         {
             Name = name;
             Description = description;
@@ -50,7 +50,7 @@ namespace POSTerminalGui
             while (line != null)
             {
                 string[] productProps = line.Split('|');
-                productList2.Add(new Product(productProps[0], productProps[1], double.Parse(productProps[2]), (Category)Enum.Parse(typeof(Category),productProps[3]), bool.Parse(productProps[4])));
+                productList2.Add(new Product(productProps[0], productProps[1], decimal.Parse(productProps[2]), (Category)Enum.Parse(typeof(Category),productProps[3]), bool.Parse(productProps[4])));
                 line = reader.ReadLine();
             }
             reader.Close();
